Beta Playtests Version 0.01
First playable version of the SRP Project.
How to Play
Keyboard and mouse is ideal (WASD and arrow keys) although you may press "X" to disable the need for a mouse.

"E" is used to interact with weapons, the staircase, and health potions. "Space" or Left Click is used to attack enemies. "Shift" is used to dash.